+++ app/guide/layout.tsx
@@ -1,5 +1,0 @@
-import MainLayout from '@/components/MainLayout';
-
-export default function Layout({ children }: { children: React.ReactNode }) {
-    return <MainLayout>{children}</MainLayout>;
-}